Убрал лишний компонент Login (уже есть SignInPage), jwt теперь сохраняется в state.
