+++ src/components/ProtectedRoute/ProtectedRoute.tsx
@@ -1,21 +1,27 @@
 import React, { useEffect } from 'react';
 import { Navigate, Outlet, useNavigate } from 'react-router-dom';
+import { rootStore } from '../../store';
 
 const ProtectedRoute: React.FC = () => {
     const navigate = useNavigate();
     const token = localStorage.getItem('token');
 
+    const {
+      profileStore: {currentUser, setCurrentUser, currentUserToken, setCurrentUserToken}
+    } = rootStore;
+
     useEffect(() => {
         const verifyToken = async () => {
-            const token = localStorage.getItem('token');
+            // const token = localStorage.getItem('token');
             try {
-                const response = await fetch(`http://localhost:8000/verify-token/${token}`);
+                const response = await fetch(`http://localhost:8000/verify-token/${currentUserToken}`);
                 if (!response.ok) {
                     throw new Error('Token verification failed');
                 }
             } catch (error) {
-                localStorage.removeItem('token');
-                navigate('/login');
+                // localStorage.removeItem('token');
+                setCurrentUserToken(null);
+                navigate('/signin');
             }
         };
         verifyToken();
