+++ src/types/user.type.ts
@@ -3,5 +3,5 @@ export default interface IUser {
   username: string,
   email: string,
   password: string,
-  roles?: Array<string>
+  role: string
 }
